requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
faiss-cpu==1.7.4
tiktoken==0.5.2
markdown-it-py==3.0.0
unstructured==0.11.0
openai>=1.10.0,<2.0.0
pypdf==3.17.0
python-docx==0.8.11
docx2txt==0.8
